Add PlantList and AddPlant components
