--- src/components/Selector.js
@@ -1,29 +1,0 @@
-import React from 'react';
-import Select from 'react-select';
-
-class Selector extends React.Component {
-    state = {
-        selectedOption: '',
-    }
-    handleChange = (selectedOption) => {
-        this.setState({ selectedOption });
-        console.log(`Selected: ${selectedOption.label}`);
-    }
-    render() {
-        const { selectedOption } = this.state;
-        const value = selectedOption && selectedOption.value;
-
-        return (
-            <Select
-                name="form-field-name"
-                value={value}
-                onChange={this.handleChange}
-                options={[
-                    { value: 'one', label: 'One' },
-                    { value: 'two', label: 'Two' },
-                ]}
-            />
-        );
-    }
-}
-export default Selector
